modelo de conceptos ahora tiene su valor

+++ models/contrato_tipo_servicio.js
@@ -24,7 +24,7 @@ module.exports = function(sequelize, DataTypes) {
       }
     },
     porcentaje: {
-      type: DataTypes.DOUBLE,
+      type: DataTypes.INTEGER,
       allowNull: true
     }
   }, {
